feat: validate attribute before saving
