+++ api/attributes/attributes.py
@@ -10,117 +10,13 @@ from ayon_server.entities import ProjectEntity
 from ayon_server.exceptions import ForbiddenException, NotFoundException
 from ayon_server.lib.postgres import Postgres
 from ayon_server.types import (
-    AttributeEnumItem,
-    AttributeType,
     OPModel,
-    ProjectLevelEntityType,
-    TopLevelEntityType,
 )
 
+from .models import AttributeModel, AttributePutModel
+from .validator import validate_attribute_data
+
 router = APIRouter(prefix="/attributes", tags=["Attributes"])
-
-
-class AttributeData(OPModel):
-    type: AttributeType = Field(
-        ...,
-        title="Type",
-        description="Type of attribute value",
-        example="string",
-    )
-    title: str | None = Field(
-        None,
-        title="Title",
-        description="Nice, human readable title of the attribute",
-        example="My attribute",
-    )
-    description: str | None = Field(
-        None,
-        title="Field description",
-        example="Value of my attribute",
-    )
-    example: Any = Field(
-        None,
-        title="Field example",
-        description="Example value of the field.",
-        example="value1",
-    )
-    default: Any = Field(
-        None,
-        title="Field default value",
-        description="Default value for the attribute. Do not set for list types.",
-    )
-    gt: int | float | None = Field(None, title="Greater than")
-    ge: int | float | None = Field(None, title="Geater or equal")
-    lt: int | float | None = Field(None, title="Less")
-    le: int | float | None = Field(None, title="Less or equal")
-    min_length: int | None = Field(None, title="Minimum length")
-    max_length: int | None = Field(None, title="Maximum length")
-    min_items: int | None = Field(
-        None,
-        title="Minimum items",
-        description="Minimum number of items in list type.",
-    )
-    max_items: int | None = Field(
-        None,
-        title="Maximum items",
-        description="Only for list types. Maximum number of items in the list.",
-    )
-    regex: str | None = Field(
-        None,
-        title="Field regex",
-        description="Only for string types. The value must match this regex.",
-        example="^[a-zA-Z0-9_]+$",
-    )
-
-    enum: list[AttributeEnumItem] | None = Field(
-        None,
-        title="Field enum",
-        description="List of enum items used for displaying select/multiselect widgets",
-        example=[
-            {"value": "value1", "label": "Value 1"},
-            {"value": "value2", "label": "Value 2"},
-            {"value": "value3", "label": "Value 3"},
-        ],
-    )
-    inherit: bool = Field(
-        True,
-        title="Inherit",
-        description="Inherit the attribute value from the parent entity.",
-    )
-
-
-class AttributeNameModel(OPModel):
-    name: str = Field(
-        ...,
-        name="Attribute name",
-        regex="^[a-zA-Z0-9]{2,30}$",
-        example="my_attribute",
-    )
-
-
-class AttributePutModel(OPModel):
-    position: int = Field(
-        ...,
-        title="Positon",
-        description="Default order",
-        example=12,
-    )
-    scope: list[ProjectLevelEntityType | TopLevelEntityType] = Field(
-        default_factory=list,
-        title="Scope",
-        description="List of entity types the attribute is available on",
-        example=["folder", "task"],
-    )
-    builtin: bool = Field(
-        False,
-        title="Builtin",
-        description="Is attribute builtin. Built-in attributes cannot be removed.",
-    )
-    data: AttributeData
-
-
-class AttributeModel(AttributePutModel, AttributeNameModel):
-    pass
 
 
 class GetAttributeListModel(OPModel):
@@ -139,7 +35,12 @@ class SetAttributeListModel(GetAttributeListModel):
     )
 
 
-async def save_attribute(attribute: AttributeModel):
+async def save_attribute(attribute: AttributeModel) -> None:
+    """Save attribute configuration to the database.
+
+    Additionally performs validation of the attribute data and updates
+    the enumerator in the running instance.
+    """
     query = """
     INSERT INTO attributes
     (name, position, scope, data)
@@ -148,6 +49,8 @@ async def save_attribute(attribute: AttributeModel):
     DO UPDATE SET position = $2, scope = $3, data = $4
     """
 
+    validate_attribute_data(attribute.name, attribute.data)
+
     await Postgres.execute(
         query,
         attribute.name,
@@ -155,6 +58,9 @@ async def save_attribute(attribute: AttributeModel):
         attribute.scope,
         attribute.data.dict(exclude_none=True),
     )
+
+    # TODO: The following code does not support horizontal scaling!!
+    # Notify other instances instead and reload the attribute library
 
     if (enum := attribute.data.enum) is not None:
         for name, field in ProjectEntity.model.attrib_model.__fields__.items():
